+++ core/components/htmlpagedom/model/htmlpagedom/htmlpagedom.class.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @package htmlpagedom
+ * @subpackage classfile
+ */
+
+require_once dirname(__DIR__, 2) . '/vendor/autoload.php';
+
+/**
+ * class HtmlPageDom
+ */
+class HtmlPageDom
+{
+}